requirements.txt
langchain-google-genai==1.0.5
langchain-text-splitters==0.2.0
langchain-pinecone==0.1.1
langchain-community==0.2.1
langchain==0.2.1
pillow==10.2.0
gitpython==3.1.43
flask==3.0.2
python-dotenv==1.0.1
pymongo[srv]==3.11
-e .